Have initCause() return an @Initialized throwable
The current behavior of returning an @UnknownInitialization throwable prevents common method chaining like: "new Throwable(e.getMessage().initCause(e)", as the initialization checker complains the result is not @Initialized. This exact pattern is suggested in the javadocs for this method.

In the absence of a @PolyInitialized annotation which seems like the more correct solution, I would proposed always marking the return value as @Initialized to allow such chaining.